backend/requirement.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
requests
sentence-transformers
chromadb
openai
